Moved iteration scope to separate file
